+++ game/Brick.h
@@ -20,7 +20,7 @@ protected:
 	int model;
 
 public:
-	Brick(int X, int Y, int W, int H, int Model);
+	Brick(float X, float Y, int W, int H, int Model);
 	virtual void Render(Camera * camera);
 	virtual void GetBoundingBox(float &l, float &t, float &r, float &b);
 	int GetModel();
