+++ game/HolyWater.cpp
@@ -104,7 +104,7 @@ void HolyWater::UpdatePositionFitSimon()
 		
 }
 
-void HolyWater::RenderIcon(int X, int Y)
+void HolyWater::RenderIcon(float X, float Y)
 {
 	_spriteIcon->Draw(X,Y);
 }
@@ -120,10 +120,10 @@ void HolyWater::Render(Camera * camera)
 
 	D3DXVECTOR2 pos = camera->Transform(x, y);
 	if (direction == -1)
-		_sprite->Draw((int)pos.x, (int)pos.y);
+		_sprite->Draw(pos.x, pos.y);
 	else
-		_sprite->DrawFlipX((int)pos.x, (int)pos.y);
-	 
+		_sprite->DrawFlipX(pos.x, pos.y);
+
 	if (IS_DEBUG_RENDER_BBOX)
 	{
  		RenderBoundingBox(camera);
