+++ game/HolyWater.h
@@ -27,7 +27,7 @@ public:
 	void GetBoundingBox(float &left, float &top, float &right, float &bottom);
 	void UpdatePositionFitSimon();
 
-	void RenderIcon(int X, int Y); 
+	void RenderIcon(float X, float Y); 
 	void Render(Camera *camera);
  
 };
