+++ game/IntroBat.cpp
@@ -35,9 +35,9 @@ void IntroBat::Render(Camera * camera)
 
 	D3DXVECTOR2 pos = camera->Transform(x, y);
 	if (direction == -1)
-		_sprite->Draw((int)pos.x, (int)pos.y);
+		_sprite->Draw(pos.x, pos.y);
 	else
-		_sprite->DrawFlipX((int)pos.x, (int)pos.y);
+		_sprite->DrawFlipX(pos.x, pos.y);
 
 	if (IS_DEBUG_RENDER_BBOX)
 		RenderBoundingBox(camera);
