+++ game/MorningStar.h
@@ -35,7 +35,7 @@ public:
 
 	void UpdatePositionFitSimon();
 	void GetBoundingBox(float &left, float &top, float &right, float &bottom);
-	void RenderIcon(int X, int Y);
+	void RenderIcon(float X, float Y);
  
 	void UpgradeLevel();
